+++ db/migrations/20180512125829_price_table.js
@@ -13,5 +13,7 @@ exports.up = function (knex, Promise) {
 }
 
 exports.down = function (knex, Promise) {
-  return knex.schema.dropTableIfExists('price')
+  if (process.env.NODE_ENV !== 'production') {
+    return knex.schema.dropTableIfExists('price')
+  }
 }
